+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - #2045 add action to input on-screen keyboard enter/send button.
 - #2106 disable the keyboard auto-switching setting when manually switching the keyboard in the Key Mapper homescreen menu.
 
+## Fixed
+
+- #2091 show an error on the "open device assistant" action when no device assistant is installed.
+
 ## [4.0.5](https://github.com/sds100/KeyMapper/releases/tag/v4.0.5)
 
 #### 26 February 2026